requirements.txt
-i https://pypi.org/simple
ansicon==1.89.0 ; platform_system == 'Windows'
arrow==1.2.3 ; python_version >= '3.6'
asgiref==3.6.0 ; python_version >= '3.7'
blessed==1.20.0 ; python_version >= '2.7'
certifi==2022.12.7 ; python_version >= '3.6'
charset-normalizer==3.0.1 ; python_version >= '3.6'
colorama==0.4.6 ; sys_platform == 'win32'
coreapi==2.3.3
coreschema==0.0.4
django==4.1.7
django-cors-headers==3.13.0
django-ebhealthcheck==2.0.2
django-filter==22.1
django-hosts==5.2
django-phonenumber-field[phonenumbers]==7.0.2
django-picklefield==3.1 ; python_version >= '3'
django-q==1.3.9
django-redis==5.2.0
django-storages==1.13.2
django-user-visit==1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
greenlet==2.0.2 ; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
gunicorn==20.1.0
hiredis==2.2.2
idna==3.4 ; python_version >= '3.5'
inflection==0.5.1 ; python_version >= '3.5'
itypes==1.2.0
jinja2==3.1.2 ; python_version >= '3.7'
jinxed==1.2.0 ; platform_system == 'Windows'
loguru==0.6.0
markupsafe==2.1.2 ; python_version >= '3.7'
numpy==1.24.2 ; python_version >= '3.10'
packaging==23.0 ; python_version >= '3.7'
pandas==1.5.3
phonenumbers==8.13.6
pillow==9.4.0
psycopg2-binary==2.9.5
pyjwt==2.6.0 ; python_version >= '3.7'
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2022.7.1
redis==3.5.3 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
requests==2.28.2
ruamel.yaml==0.17.21 ; python_version >= '3'
ruamel.yaml.clib==0.2.7 ; python_version < '3.11' and platform_python_implementation == 'CPython'
scipy==1.10.0
setuptools==67.3.2 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==2.0.3
sqlparse==0.4.3 ; python_version >= '3.5'
typing-extensions==4.5.0 ; python_version >= '3.7'
tzdata==2022.7 ; sys_platform == 'win32'
ua-parser==0.16.1
uritemplate==4.1.1 ; python_version >= '3.6'
urllib3==1.26.14 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
user-agents==2.2.0
wcwidth==0.2.6
win32-setctime==1.1.0 ; sys_platform == 'win32'
